FileTransfer v.0.5.1
Hot fix: Fixed the "bad decrypt" bug that was noticed when restarting and re-downloading the file. Now the encryption/decryption keys are stored next to the downloaded file in the "files" directory.
Changes: Now the original name is returned to the files when downloading (instead of the uuid that is assigned when uploading).
Added: The ability to set a dark/light theme, and it is saved in the browser. If you are using the nginx proxy server, now the display of client ip addresses will be displayed correctly, since support for headers has been added.